Temporary fix for inversed source and destination address fields on network packets.
More information:

While re-implementing monitor to use aya-rs instead of redbpf, I chose
to use `xdp` instead of `socket_filter` for the potential performance gain.

But since `xdp` is attached to the virtual ethernet of a container, this
caused the source and destination fields of packets being inverted.

The correct fix should be to switch `xdp` for `socket_filter`.