Update communication and emulation modules from the emulationcore to run as separate tokio tasks and use message passing.
communication now contains all capnp related logic for sending/receiving metadata, making it easier to use in general.

Both can no be shared as they are cheap to clone thanks to the underlying mpsc::Sender.

Also since emulationcore already uses a tokio runtime, update it to use the async equivalent of std's thread::sleep, sync::Mutex, and net::TcpListener, so that it can yield control of the thread and advance any other tasks instead of blocking it.

And lastly bump the cargo workspace resolver to version 3.